refactor: move tournament team to its own file
