+++ frontend/src/types/team.ts
@@ -4,10 +4,3 @@ export interface TeamSummary {
   short_name: string;
   logo_url: string | null;
 }
-
-export interface TournamentTeam {
-  summary: TeamSummary;
-  group: string | null;
-  final_rank: number | null;
-  stage_reached: string | null;
-}
